MiServices: Introduce Quickspace
Quickspace is a service that presents useful information to the 
user such as Weather, Alarms and Calendar Events as well as useful
tips on the Lockscreen and Homescreen. Quickspace is based on the
Smartspace concept created by Google.

Weather logic is based on all components in 
https://github.com/CypherOS/packages_providers_WeatherProvider with
credits to the original dev @jhenrique09 and main contributor @ezio84

Previous logic handled queries in framework but struggled with multiple
queries from different packages, this makes that process simplier with
portability across Android versions. In this commit, Quickspace supports
weather and 1 QuickEvent (Device introduction) with more QuickEvents being
added later on

Change-Id: I0bdeb1b743c3f019588523413ac1e97f1d3550a1